Export needed bootstrap methods and variables
Also add back validating persisted config and update tests.
